pyproject.toml
[project]
name = "blender_console_mirror"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = []

[dependency-groups]
dev = [
    "fake-bpy-module-latest>=20251003",
    "mypy>=1.18.2",
    "pytest>=9.0.1",
    "ruff>=0.14.5",
]

[tool.pytest.ini_options]
testpaths = ["tests/unit"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = ["blender_console_mirror"]
